backend/EvaluateSurveyService/requirements.txt
google-cloud-pubsub==2.9.0
firebase-admin==5.2.0